fix: data storage and math

+++ types/team.go
@@ -1,9 +1,9 @@
 package types
 
 type Team struct {
-	LastUpdated string `json:"last_updated"`
-	Name        string `json:"name"`
-	Tag         string `json:"tag"`
-	Rating      int    `json:"rating"`
-	Games       []Game `json:"games"`
+	LastUpdated string  `json:"last_updated"`
+	Name        string  `json:"name"`
+	Tag         string  `json:"tag"`
+	Rating      float64 `json:"rating"`
+	Games       []Game  `json:"games"`
 }
